+++ internal/pg_documentdb_rum/documentdb_rum--1.0.sql
@@ -1,0 +1,10 @@
+CREATE FUNCTION documentdb_rumhandler(internal)
+RETURNS index_am_handler
+AS 'MODULE_PATHNAME'
+LANGUAGE C;
+
+/*
+ * RUM access method
+ */
+
+CREATE ACCESS METHOD documentdb_rum TYPE INDEX HANDLER documentdb_rumhandler;